src/main/java/com/shcherbinina/cinemapark/validation/payloadValidation/AccountValidator.java
package com.shcherbinina.cinemapark.validation.payloadValidation;

import org.springframework.stereotype.Component;
import org.springframework.validation.Errors;
import org.springframework.validation.Validator;

@Component
public class AccountValidator implements Validator {
    private final String invalidAmountOfMoney = "Invalid amount of money";

    @Override
    public boolean supports(Class<?> aClass) {
        return false;
    }

    @Override
    public void validate(Object o, Errors errors) {

    }
}
